requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
packaging==24.0
pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
tk==0.1.0
tkhtmlview==0.3.0
ttkbootstrap==1.10.1
ttkthemes==3.2.2
urllib3==2.2.1
